requirements.txt
datasets==2.10.0
transformers==4.26.1
numpy==1.24.2
joblib==1.2.0
#chardet==5.1.0
cchardet == 2.1.7
scikit-learn== 1.2.1
pandas==1.5.3
tqdm==4.64.1
tensorboard==2.12.0
accelerate == 0.16.0
torch-tb-profiler== 0.4.1
evaluate==0.4.0
huggingface_hub == 0.13.2